requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0
pyperclip>=1.8.0
pystray>=0.19.0
Pillow>=10.3.0
qrcode>=7.4.0
zeroconf>=0.131.0
python-multipart>=0.0.9
aiofiles>=23.2.0
